# webapp/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pok\u00e9mon TCG Companion",
    "version": "0.1.0"
  },
  "paths": {
    "/cards/upload-url": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Create Upload Url",
        "operationId": "create_upload_url_cards_upload_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardSignedUploadUrl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "List Cards",
        "operationId": "list_cards_cards_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CardRead"
                  },
                  "type": "array",
                  "title": "Response List Cards Cards Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Add Card",
        "operationId": "add_card_cards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card_id}": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "Get Card",
        "operationId": "get_card_cards__card_id__get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/webhooks/card-created": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Handle Card Created",
        "operationId": "handle_card_created_cards_webhooks_card_created_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PubSubEnvelope"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Card": {
        "properties": {
          "ref_card_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Card Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "image_path"
        ],
        "title": "Card"
      },
      "CardRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ref_card_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Card Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path"
          },
          "ref_card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RefCardRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "image_path"
        ],
        "title": "CardRead"
      },
      "CreateCardPayload": {
        "properties": {
          "image_path": {
            "type": "string",
            "title": "Image Path"
          }
        },
        "type": "object",
        "required": [
          "image_path"
        ],
        "title": "CreateCardPayload"
      },
      "CreateCardSignedUploadUrl": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content_type"
        ],
        "title": "CreateCardSignedUploadUrl"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PubSubEnvelope": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/PubSubMessage"
          },
          "subscription": {
            "type": "string",
            "title": "Subscription"
          }
        },
        "type": "object",
        "required": [
          "message",
          "subscription"
        ],
        "title": "PubSubEnvelope",
        "description": "Pub/Sub push subscription envelope."
      },
      "PubSubMessage": {
        "properties": {
          "data": {
            "type": "string",
            "title": "Data"
          },
          "messageId": {
            "type": "string",
            "title": "Messageid"
          },
          "publishTime": {
            "type": "string",
            "title": "Publishtime"
          }
        },
        "type": "object",
        "required": [
          "data",
          "messageId",
          "publishTime"
        ],
        "title": "PubSubMessage",
        "description": "Pub/Sub message structure."
      },
      "RefCardRead": {
        "properties": {
          "tcg_id": {
            "type": "string",
            "title": "Tcg Id"
          },
          "tcg_local_id": {
            "type": "string",
            "title": "Tcg Local Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "set_id": {
            "type": "string",
            "title": "Set Id"
          },
          "set_name": {
            "type": "string",
            "title": "Set Name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "tcg_id",
          "tcg_local_id",
          "name",
          "set_id",
          "set_name",
          "id"
        ],
        "title": "RefCardRead"
      },
      "UploadUrlResponse": {
        "properties": {
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path"
          }
        },
        "type": "object",
        "required": [
          "upload_url",
          "image_path"
        ],
        "title": "UploadUrlResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}